+++ src/main.ts
@@ -1,5 +1,7 @@
 import { Command } from "commander";
-import { handleAction} from "./user";
+import { getUserName, fnv1aHash} from "./user";
+import { llmNames } from "./llms";
+import { startChat, askForModelChoice } from "./chat";
 
 const program = new Command();
 
@@ -8,9 +10,34 @@ program
   .description('CLI Chat Application with LLM')
   .option('-n, --name <name>', 'Specify your name')
   .option('-m, --model <model>', 'Choose the model to use')
-  .action(handleAction);
+  .action(handleChat);
 
 program.parse(process.argv);
+
+export async function handleChat(options: any): Promise<void> {
+  let userName = options.name || await getUserName();
+  let userHash = fnv1aHash(userName)
+  console.log(`FNV-1a Hash for ${userName}: ${userHash}`);
+
+  let model = options.model || await askForModelChoice();
+  let modelName = llmNames.find(llmName => llmName.toLowerCase() === model?.toLowerCase());
+
+  if (!modelName) {
+    console.log(`Invalid model. Please choose a valid model.`);
+    model = await askForModelChoice();
+    modelName = llmNames.find(llmName => llmName.toLowerCase() === model?.toLowerCase());
+  }
+
+  console.log(`Hello, ${userName}! Welcome to the CLI application.`);
+  console.log(`Model chosen: ${modelName}`);
+
+  if (userName && modelName) {
+    console.log(`Processing to chat...`);
+    await startChat(modelName);
+  } else {
+    console.error('Error: Missing user name or invalid model choice.');
+  }
+}
 
 function handleExit() {
   process.exit();
